+++ McDermott/McDermott.Application/Dtos/Config/DistrictDto.cs
@@ -2,16 +2,16 @@
 {
     public class DistrictDto : IMapFrom<District>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Required]
         [StringLength(200)]
         public string Name { get; set; } = string.Empty; // Kecamatan
         [Required]
-        public int? CityId { get; set; } // Kabupaten
+        public long? CityId { get; set; } // Kabupaten
         [Required]
-        public int? ProvinceId { get; set; }
-       
+        public long? ProvinceId { get; set; }
+
         public virtual CityDto? City { get; set; }
         public virtual ProvinceDto? Province { get; set; }
     }
