+++ McDermott/McDermott.Application/Dtos/Queue/KioskDto.cs
@@ -2,7 +2,7 @@
 {
     public partial class KioskDto : IMapFrom<Kiosk>
     {
-        public int Id { get; set; }
+         public long Id { get; set; }
         public string? Type { get; set; }
 
         [Required]
@@ -10,13 +10,13 @@
         [RegularExpression(@"^\d+$", ErrorMessage = "The {0} field must contain only numbers.")]
         public string? NumberType { get; set; } = string.Empty;
 
-        public int? PatientId { get; set; }
+        public long? PatientId { get; set; }
         public string? BPJS { get; set; } = string.Empty;
         public bool? StageBpjs { get; set; }
-        public int? ServiceId { get; set; }
-        public int? PhysicianId { get; set; }
-        public int? Queue { get; set; }
-        public int? CounterId { get; set; }
+        public long? ServiceId { get; set; }
+        public long? PhysicianId { get; set; }
+        public long? Queue { get; set; }
+        public long? CounterId { get; set; }
         public virtual CounterDto? Counter { get; set; }
         public virtual ServiceDto? Service { get; set; }
         public virtual UserDto? Patient { get; set; }
